Enable API and tests to execute using Docker
Both the Node.js API and the Rspec tests now run within their
own separate Docker containers. Docker-Compose is used to run both
containers and link API to the test container.

Both test and API code can be modified without requiring a rebuild
of a container. A rebuild of the test container will be triggered
if any of the dependencies change within the Gemfile.